# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ShortMesh Authy service",
        "title": "Shortmesh - Authy API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/users.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/users.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/users.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/users.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Create a new user account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/users.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/users.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "User with email already exists",
                        "schema": {
                            "$ref": "#/definitions/users.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/users.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/otp/generate": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Generate a secure OTP code and send it to the specified identifier via the interface API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "otp"
                ],
                "summary": "Generate and send OTP",
                "parameters": [
                    {
                        "description": "OTP generation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/otp.GenerateOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "schema": {
                            "$ref": "#/definitions/otp.GenerateOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/otp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/otp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/otp/verify": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Verify an OTP code for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "otp"
                ],
                "summary": "Verify OTP",
                "parameters": [
                    {
                        "description": "OTP verification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/otp.VerifyOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP verified successfully",
                        "schema": {
                            "$ref": "#/definitions/otp.VerifyOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/otp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/otp.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "TooManyRequests",
                        "schema": {
                            "$ref": "#/definitions/otp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/otp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/platforms": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "List all available platforms and their senders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platforms"
                ],
                "summary": "List available platforms",
                "responses": {
                    "200": {
                        "description": "List of platforms retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/platforms.Platform"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/platforms.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/platforms.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "otp.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "otp.GenerateOTPRequest": {
            "type": "object",
            "required": [
                "identifier",
                "platform",
                "sender"
            ],
            "properties": {
                "identifier": {
                    "type": "string",
                    "example": "237123456780"
                },
                "platform": {
                    "type": "string",
                    "example": "wa"
                },
                "sender": {
                    "type": "string",
                    "example": "237123456789"
                }
            }
        },
        "otp.GenerateOTPResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2026-02-19T20:30:00Z"
                },
                "message": {
                    "type": "string",
                    "example": "OTP sent successfully"
                }
            }
        },
        "otp.VerifyOTPRequest": {
            "type": "object",
            "required": [
                "code",
                "identifier",
                "platform",
                "sender"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "identifier": {
                    "type": "string",
                    "example": "237123456780"
                },
                "platform": {
                    "type": "string",
                    "example": "wa"
                },
                "sender": {
                    "type": "string",
                    "example": "237123456789"
                }
            }
        },
        "otp.VerifyOTPResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "OTP verified successfully"
                }
            }
        },
        "platforms.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "platforms.Platform": {
            "type": "object",
            "properties": {
                "platform": {
                    "type": "string",
                    "example": "wa"
                },
                "sender": {
                    "type": "string",
                    "example": "237123456789"
                }
            }
        },
        "users.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8,
                    "example": "Validpassword123@!"
                }
            }
        },
        "users.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "users.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "Validpassword123@!"
                }
            }
        },
        "users.UserResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "CYDD4YLEmwQ3TCiSFidaFQ=="
                },
                "client_secret": {
                    "type": "string",
                    "example": "ux5Wa2P790vyHTzeaKWQkpfUb0BNGA7qQkz5JnH4k-Y="
                },
                "message": {
                    "type": "string",
                    "example": "User created successfully"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}